Clean up and enhancement:  - documentation  - fixes and more clarified code
cooperative mode:
old way: if one BlockBase matcher is in process, no other matchers can
  be in effect
new way: if in cooperative mode, all matchers can be at work. This is 
  the default. If non-cooperative, it behaves as it used to.

Added a _processLine() hook in the BlockBase so subclasses have a 
  chance to operate on the line.

Added more BlockMatchex’es. One to find a particular match, and then
  transform to a regex to get all the matches. e.g. the “\s(\d\d)\s”
  expression would normally just return the first occurrence of 2
  digits, but we can do two passes and get all the occurrences with 
  the PatternMatchex.

Made the URLStreamParser a subclass of a StreamContext, so we can
know which lines have been parsed and the line numbers.