Huge reorganization to simplify the code.
The major changes are in BlockBase. Reduced the amount of state, so that we
are calculating things. The only bit of state is the regex index, which only
covers the preceding regex'es to the main matching regex. The state (the
constants in the beginning of the file) is computed based on that.

Parser was consolidated to reuse a little code in the parse() function.

flake8 checks. It's not 100%, but much closer.

tests: not nearly comprehensive, but covers some basic use cases. I'm not
actually sure the original use (Xcode project files) even works anymore.
The test exists, but it was just broken out of the BlockBase file and not
really made into a real unittest.TestCase.